requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
langchain-pinecone>=0.1.0
langchain-text-splitters>=0.0.1
openai>=1.3.0
pinecone-client>=3.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.2
requests>=2.31.0
numpy>=1.24.0
google-generativeai>=0.3.0